requirements.txt
hydra-core
litellm
numpy
omegaconf
pydantic
PyYAML
scipy
tqdm
tiktoken
numba
networkx
joblib
requests
flask
torch
transformers
radon
autopep8
python-dotenv
scikit-learn